Support complex arrays

+++ skfem/utils.py
@@ -657,7 +657,7 @@ def mpc(A: spmatrix,
     return (
         B,
         y,
-        np.zeros_like(b),
+        np.zeros_like(b, dtype=B.dtype),
         (
             np.concatenate((U, M, S)),
             lambda x: np.concatenate((x, T @ x[len(U):] + g)),
